requirements.txt
streamlit
pandas
plotly
openai
python-dotenv
langchain
langchain-community
langchain-openai
faiss-cpu
